+++ PHPUnit/Framework/Assert/Functions.php
@@ -37,7 +37,6 @@
  * @package    PHPUnit
  * @subpackage Framework
  * @author     Sebastian Bergmann <sebastian@phpunit.de>
- * @author     Kuzuha SHINODA <kuzuha01@hotmail.com>
  * @copyright  2001-2013 Sebastian Bergmann <sebastian@phpunit.de>
  * @license    http://www.opensource.org/licenses/BSD-3-Clause  The BSD 3-Clause License
  * @link       http://www.phpunit.de/
@@ -53,7 +52,10 @@
  */
 function any()
 {
-    return PHPUnit_Framework_TestCase::any();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::any',
+      func_get_args()
+    );
 }
 
 /**
@@ -64,7 +66,10 @@ function any()
  */
 function anything()
 {
-    return PHPUnit_Framework_Assert::anything();
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::anything',
+      func_get_args()
+    );
 }
 
 /**
@@ -76,33 +81,42 @@ function anything()
  */
 function arrayHasKey($key)
 {
-    return PHPUnit_Framework_Assert::arrayHasKey($key);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::arrayHasKey',
+      func_get_args()
+    );
 }
 
 /**
  * Asserts that an array has a specified key.
  *
  * @param  mixed  $key
- * @param  array  $array
+ * @param  array|ArrayAccess  $array
  * @param  string $message
  * @since  Method available since Release 3.0.0
  */
-function assertArrayHasKey($key, array $array, $message = '')
+function assertArrayHasKey($key, $array, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertArrayHasKey($key, $array, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertArrayHasKey',
+      func_get_args()
+    );
 }
 
 /**
  * Asserts that an array does not have a specified key.
  *
  * @param  mixed  $key
- * @param  array  $array
+ * @param  array|ArrayAccess  $array
  * @param  string $message
  * @since  Method available since Release 3.0.0
  */
-function assertArrayNotHasKey($key, array $array, $message = '')
+function assertArrayNotHasKey($key, $array, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertArrayNotHasKey($key, $array, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertArrayNotHasKey',
+      func_get_args()
+    );
 }
 
 /**
@@ -119,7 +133,10 @@ function assertArrayNotHasKey($key, array $array, $message = '')
  */
 function assertAttributeContains($needle, $haystackAttributeName, $haystackClassOrObject, $message = '', $ignoreCase = FALSE, $checkForObjectIdentity = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertAttributeContains($needle, $haystackAttributeName, $haystackClassOrObject, $message, $ignoreCase, $checkForObjectIdentity);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeContains',
+      func_get_args()
+    );
 }
 
 /**
@@ -135,7 +152,10 @@ function assertAttributeContains($needle, $haystackAttributeName, $haystackClass
  */
 function assertAttributeContainsOnly($type, $haystackAttributeName, $haystackClassOrObject, $isNativeType = NULL, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeContainsOnly($type, $haystackAttributeName, $haystackClassOrObject, $isNativeType, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeContainsOnly',
+      func_get_args()
+    );
 }
 
 /**
@@ -150,7 +170,10 @@ function assertAttributeContainsOnly($type, $haystackAttributeName, $haystackCla
  */
 function assertAttributeCount($expectedCount, $haystackAttributeName, $haystackClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeCount($expectedCount, $haystackAttributeName, $haystackClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeCount',
+      func_get_args()
+    );
 }
 
 /**
@@ -164,7 +187,10 @@ function assertAttributeCount($expectedCount, $haystackAttributeName, $haystackC
  */
 function assertAttributeEmpty($haystackAttributeName, $haystackClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeEmpty($haystackAttributeName, $haystackClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeEmpty',
+      func_get_args()
+    );
 }
 
 /**
@@ -181,7 +207,10 @@ function assertAttributeEmpty($haystackAttributeName, $haystackClassOrObject, $m
  */
 function assertAttributeEquals($expected, $actualAttributeName, $actualClassOrObject, $message = '', $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertAttributeEquals($expected, $actualAttributeName, $actualClassOrObject, $message, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -195,7 +224,10 @@ function assertAttributeEquals($expected, $actualAttributeName, $actualClassOrOb
  */
 function assertAttributeGreaterThan($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeGreaterThan($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeGreaterThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -209,7 +241,10 @@ function assertAttributeGreaterThan($expected, $actualAttributeName, $actualClas
  */
 function assertAttributeGreaterThanOrEqual($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeGreaterThanOrEqual($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeGreaterThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -223,7 +258,10 @@ function assertAttributeGreaterThanOrEqual($expected, $actualAttributeName, $act
  */
 function assertAttributeInstanceOf($expected, $attributeName, $classOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeInstanceOf($expected, $attributeName, $classOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeInstanceOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -237,7 +275,10 @@ function assertAttributeInstanceOf($expected, $attributeName, $classOrObject, $m
  */
 function assertAttributeInternalType($expected, $attributeName, $classOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeInternalType($expected, $attributeName, $classOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeInternalType',
+      func_get_args()
+    );
 }
 
 /**
@@ -251,7 +292,10 @@ function assertAttributeInternalType($expected, $attributeName, $classOrObject, 
  */
 function assertAttributeLessThan($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeLessThan($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeLessThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -265,7 +309,10 @@ function assertAttributeLessThan($expected, $actualAttributeName, $actualClassOr
  */
 function assertAttributeLessThanOrEqual($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeLessThanOrEqual($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeLessThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -282,7 +329,10 @@ function assertAttributeLessThanOrEqual($expected, $actualAttributeName, $actual
  */
 function assertAttributeNotContains($needle, $haystackAttributeName, $haystackClassOrObject, $message = '', $ignoreCase = FALSE, $checkForObjectIdentity = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotContains($needle, $haystackAttributeName, $haystackClassOrObject, $message, $ignoreCase, $checkForObjectIdentity);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotContains',
+      func_get_args()
+    );
 }
 
 /**
@@ -299,7 +349,10 @@ function assertAttributeNotContains($needle, $haystackAttributeName, $haystackCl
  */
 function assertAttributeNotContainsOnly($type, $haystackAttributeName, $haystackClassOrObject, $isNativeType = NULL, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotContainsOnly($type, $haystackAttributeName, $haystackClassOrObject, $isNativeType, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotContainsOnly',
+      func_get_args()
+    );
 }
 
 /**
@@ -314,7 +367,10 @@ function assertAttributeNotContainsOnly($type, $haystackAttributeName, $haystack
  */
 function assertAttributeNotCount($expectedCount, $haystackAttributeName, $haystackClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotCount($expectedCount, $haystackAttributeName, $haystackClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotCount',
+      func_get_args()
+    );
 }
 
 /**
@@ -328,7 +384,10 @@ function assertAttributeNotCount($expectedCount, $haystackAttributeName, $haysta
  */
 function assertAttributeNotEmpty($haystackAttributeName, $haystackClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotEmpty($haystackAttributeName, $haystackClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotEmpty',
+      func_get_args()
+    );
 }
 
 /**
@@ -345,7 +404,10 @@ function assertAttributeNotEmpty($haystackAttributeName, $haystackClassOrObject,
  */
 function assertAttributeNotEquals($expected, $actualAttributeName, $actualClassOrObject, $message = '', $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotEquals($expected, $actualAttributeName, $actualClassOrObject, $message, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -359,7 +421,10 @@ function assertAttributeNotEquals($expected, $actualAttributeName, $actualClassO
  */
 function assertAttributeNotInstanceOf($expected, $attributeName, $classOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotInstanceOf($expected, $attributeName, $classOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotInstanceOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -373,7 +438,10 @@ function assertAttributeNotInstanceOf($expected, $attributeName, $classOrObject,
  */
 function assertAttributeNotInternalType($expected, $attributeName, $classOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotInternalType($expected, $attributeName, $classOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotInternalType',
+      func_get_args()
+    );
 }
 
 /**
@@ -387,7 +455,10 @@ function assertAttributeNotInternalType($expected, $attributeName, $classOrObjec
  */
 function assertAttributeNotSame($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeNotSame($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeNotSame',
+      func_get_args()
+    );
 }
 
 /**
@@ -401,7 +472,10 @@ function assertAttributeNotSame($expected, $actualAttributeName, $actualClassOrO
  */
 function assertAttributeSame($expected, $actualAttributeName, $actualClassOrObject, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertAttributeSame($expected, $actualAttributeName, $actualClassOrObject, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertAttributeSame',
+      func_get_args()
+    );
 }
 
 /**
@@ -414,7 +488,10 @@ function assertAttributeSame($expected, $actualAttributeName, $actualClassOrObje
  */
 function assertClassHasAttribute($attributeName, $className, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertClassHasAttribute($attributeName, $className, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertClassHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -427,7 +504,10 @@ function assertClassHasAttribute($attributeName, $className, $message = '')
  */
 function assertClassHasStaticAttribute($attributeName, $className, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertClassHasStaticAttribute($attributeName, $className, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertClassHasStaticAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -440,7 +520,10 @@ function assertClassHasStaticAttribute($attributeName, $className, $message = ''
  */
 function assertClassNotHasAttribute($attributeName, $className, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertClassNotHasAttribute($attributeName, $className, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertClassNotHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -453,7 +536,10 @@ function assertClassNotHasAttribute($attributeName, $className, $message = '')
  */
 function assertClassNotHasStaticAttribute($attributeName, $className, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertClassNotHasStaticAttribute($attributeName, $className, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertClassNotHasStaticAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -468,7 +554,10 @@ function assertClassNotHasStaticAttribute($attributeName, $className, $message =
  */
 function assertContains($needle, $haystack, $message = '', $ignoreCase = FALSE, $checkForObjectIdentity = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertContains($needle, $haystack, $message, $ignoreCase, $checkForObjectIdentity);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertContains',
+      func_get_args()
+    );
 }
 
 /**
@@ -482,7 +571,25 @@ function assertContains($needle, $haystack, $message = '', $ignoreCase = FALSE, 
  */
 function assertContainsOnly($type, $haystack, $isNativeType = NULL, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertContainsOnly($type, $haystack, $isNativeType, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertContainsOnly',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that a haystack contains only instances of a given classname
+ *
+ * @param string $classname
+ * @param array|Traversable $haystack
+ * @param string $message
+ */
+function assertContainsOnlyInstancesOf($classname, $haystack, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertContainsOnlyInstancesOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -494,7 +601,10 @@ function assertContainsOnly($type, $haystack, $isNativeType = NULL, $message = '
  */
 function assertCount($expectedCount, $haystack, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertCount($expectedCount, $haystack, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertCount',
+      func_get_args()
+    );
 }
 
 /**
@@ -506,7 +616,10 @@ function assertCount($expectedCount, $haystack, $message = '')
  */
 function assertEmpty($actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertEmpty($actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertEmpty',
+      func_get_args()
+    );
 }
 
 /**
@@ -521,7 +634,10 @@ function assertEmpty($actual, $message = '')
  */
 function assertEqualXMLStructure(DOMElement $expectedElement, DOMElement $actualElement, $checkAttributes = FALSE, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertEqualXMLStructure($expectedElement, $actualElement, $checkAttributes, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertEqualXMLStructure',
+      func_get_args()
+    );
 }
 
 /**
@@ -537,7 +653,10 @@ function assertEqualXMLStructure(DOMElement $expectedElement, DOMElement $actual
  */
 function assertEquals($expected, $actual, $message = '', $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertEquals($expected, $actual, $message, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -549,7 +668,10 @@ function assertEquals($expected, $actual, $message = '', $delta = 0, $maxDepth =
  */
 function assertFalse($condition, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertFalse($condition, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertFalse',
+      func_get_args()
+    );
 }
 
 /**
@@ -565,7 +687,10 @@ function assertFalse($condition, $message = '')
  */
 function assertFileEquals($expected, $actual, $message = '', $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertFileEquals($expected, $actual, $message, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertFileEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -577,7 +702,10 @@ function assertFileEquals($expected, $actual, $message = '', $canonicalize = FAL
  */
 function assertFileExists($filename, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertFileExists($filename, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertFileExists',
+      func_get_args()
+    );
 }
 
 /**
@@ -593,7 +721,10 @@ function assertFileExists($filename, $message = '')
  */
 function assertFileNotEquals($expected, $actual, $message = '', $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertFileNotEquals($expected, $actual, $message, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertFileNotEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -605,7 +736,10 @@ function assertFileNotEquals($expected, $actual, $message = '', $canonicalize = 
  */
 function assertFileNotExists($filename, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertFileNotExists($filename, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertFileNotExists',
+      func_get_args()
+    );
 }
 
 /**
@@ -618,7 +752,10 @@ function assertFileNotExists($filename, $message = '')
  */
 function assertGreaterThan($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertGreaterThan($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertGreaterThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -631,7 +768,10 @@ function assertGreaterThan($expected, $actual, $message = '')
  */
 function assertGreaterThanOrEqual($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertGreaterThanOrEqual($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertGreaterThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -644,7 +784,10 @@ function assertGreaterThanOrEqual($expected, $actual, $message = '')
  */
 function assertInstanceOf($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertInstanceOf($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertInstanceOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -657,72 +800,10 @@ function assertInstanceOf($expected, $actual, $message = '')
  */
 function assertInternalType($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertInternalType($expected, $actual, $message);
-}
-
-/**
- * Asserts that two given JSON encoded objects or arrays are equal.
- *
- * @param string $expectedJson
- * @param string $actualJson
- * @param string $message
- * @since Method available since Release 3.7.0
- */
-function assertJsonStringEqualsJsonString($expectedJson, $actualJson, $message = '')
-{
-    return PHPUnit_Framework_Assert::assertJsonStringEqualsJsonString($expectedJson, $actualJson, $message = '');
-}
-
-/**
- * Asserts that two given JSON encoded objects or arrays are not equal.
- *
- * @param string $expectedJson
- * @param string $actualJson
- * @param string $message
- * @since Method available since Release 3.7.0
- */
-function assertJsonStringNotEqualsJsonString($expectedJson, $actualJson, $message = '')
-{
-    return PHPUnit_Framework_Assert::assertJsonStringNotEqualsJsonString($expectedJson, $actualJson, $message = '');
-}
-
-/**
- * Asserts that the generated JSON encoded object and the content of the given file are equal.
- *
- * @param string $expectedFile
- * @param string $actualJson
- * @param string $message
- * @since Method available since Release 3.7.0
- */
-function assertJsonStringEqualsJsonFile($expectedFile, $actualJson, $message = '')
-{
-    return PHPUnit_Framework_Assert::assertJsonStringEqualsJsonFile($expectedFile, $actualJson, $message = '');
-}
-
-/**
- * Asserts that the generated JSON encoded object and the content of the given file are not equal.
- *
- * @param string $expectedFile
- * @param string $actualJson
- * @param string $message
- * @since Method available since Release 3.7.0
- */
-function assertJsonStringNotEqualsJsonFile($expectedFile, $actualJson, $message = '')
-{
-    return PHPUnit_Framework_Assert::assertJsonStringNotEqualsJsonFile($expectedFile, $actualJson, $message = '');
-}
-
-/**
- * Asserts that two JSON files are not equal.
- *
- * @param  string $expectedFile
- * @param  string $actualFile
- * @param  string $message
- * @since Method available since Release 3.7.0
- */
-function assertJsonFileNotEqualsJsonFile($expectedFile, $actualFile, $message = '')
-{
-    return PHPUnit_Framework_Assert::assertJsonFileNotEqualsJsonFile($expectedFile, $actualFile, $message = '');
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertInternalType',
+      func_get_args()
+    );
 }
 
 /**
@@ -731,11 +812,88 @@ function assertJsonFileNotEqualsJsonFile($expectedFile, $actualFile, $message = 
  * @param  string $expectedFile
  * @param  string $actualFile
  * @param  string $message
- * @since Method available since Release 3.7.0
  */
 function assertJsonFileEqualsJsonFile($expectedFile, $actualFile, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertJsonFileEqualsJsonFile($expectedFile, $actualFile, $message = '');
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonFileEqualsJsonFile',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that two JSON files are not equal.
+ *
+ * @param  string $expectedFile
+ * @param  string $actualFile
+ * @param  string $message
+ */
+function assertJsonFileNotEqualsJsonFile($expectedFile, $actualFile, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonFileNotEqualsJsonFile',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that the generated JSON encoded object and the content of the given file are equal.
+ *
+ * @param string $expectedFile
+ * @param string $actualJson
+ * @param string $message
+ */
+function assertJsonStringEqualsJsonFile($expectedFile, $actualJson, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonStringEqualsJsonFile',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that two given JSON encoded objects or arrays are equal.
+ *
+ * @param string $expectedJson
+ * @param string $actualJson
+ * @param string $message
+ */
+function assertJsonStringEqualsJsonString($expectedJson, $actualJson, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonStringEqualsJsonString',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that the generated JSON encoded object and the content of the given file are not equal.
+ *
+ * @param string $expectedFile
+ * @param string $actualJson
+ * @param string $message
+ */
+function assertJsonStringNotEqualsJsonFile($expectedFile, $actualJson, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonStringNotEqualsJsonFile',
+      func_get_args()
+    );
+}
+
+/**
+ * Asserts that two given JSON encoded objects or arrays are not equal.
+ *
+ * @param string $expectedJson
+ * @param string $actualJson
+ * @param string $message
+ */
+function assertJsonStringNotEqualsJsonString($expectedJson, $actualJson, $message = '')
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertJsonStringNotEqualsJsonString',
+      func_get_args()
+    );
 }
 
 /**
@@ -748,7 +906,10 @@ function assertJsonFileEqualsJsonFile($expectedFile, $actualFile, $message = '')
  */
 function assertLessThan($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertLessThan($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertLessThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -761,7 +922,10 @@ function assertLessThan($expected, $actual, $message = '')
  */
 function assertLessThanOrEqual($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertLessThanOrEqual($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertLessThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -776,7 +940,10 @@ function assertLessThanOrEqual($expected, $actual, $message = '')
  */
 function assertNotContains($needle, $haystack, $message = '', $ignoreCase = FALSE, $checkForObjectIdentity = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertNotContains($needle, $haystack, $message, $ignoreCase, $checkForObjectIdentity);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotContains',
+      func_get_args()
+    );
 }
 
 /**
@@ -790,7 +957,10 @@ function assertNotContains($needle, $haystack, $message = '', $ignoreCase = FALS
  */
 function assertNotContainsOnly($type, $haystack, $isNativeType = NULL, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotContainsOnly($type, $haystack, $isNativeType, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotContainsOnly',
+      func_get_args()
+    );
 }
 
 /**
@@ -802,7 +972,10 @@ function assertNotContainsOnly($type, $haystack, $isNativeType = NULL, $message 
  */
 function assertNotCount($expectedCount, $haystack, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotCount($expectedCount, $haystack, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotCount',
+      func_get_args()
+    );
 }
 
 /**
@@ -814,7 +987,10 @@ function assertNotCount($expectedCount, $haystack, $message = '')
  */
 function assertNotEmpty($actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotEmpty($actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotEmpty',
+      func_get_args()
+    );
 }
 
 /**
@@ -831,7 +1007,10 @@ function assertNotEmpty($actual, $message = '')
  */
 function assertNotEquals($expected, $actual, $message = '', $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertNotEquals($expected, $actual, $message, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -844,7 +1023,10 @@ function assertNotEquals($expected, $actual, $message = '', $delta = 0, $maxDept
  */
 function assertNotInstanceOf($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotInstanceOf($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotInstanceOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -857,7 +1039,10 @@ function assertNotInstanceOf($expected, $actual, $message = '')
  */
 function assertNotInternalType($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotInternalType($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotInternalType',
+      func_get_args()
+    );
 }
 
 /**
@@ -868,7 +1053,10 @@ function assertNotInternalType($expected, $actual, $message = '')
  */
 function assertNotNull($actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotNull($actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotNull',
+      func_get_args()
+    );
 }
 
 /**
@@ -881,7 +1069,10 @@ function assertNotNull($actual, $message = '')
  */
 function assertNotRegExp($pattern, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotRegExp($pattern, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotRegExp',
+      func_get_args()
+    );
 }
 
 /**
@@ -895,20 +1086,26 @@ function assertNotRegExp($pattern, $string, $message = '')
  */
 function assertNotSame($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotSame($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotSame',
+      func_get_args()
+    );
 }
 
 /**
  * Assert that the size of two arrays (or `Countable` or `Iterator` objects)
  * is not the same.
  *
- * @param integer $expected
- * @param mixed   $actual
- * @param string  $message
+ * @param array|Countable|Iterator $expected
+ * @param array|Countable|Iterator $actual
+ * @param string $message
  */
-function assertNotSameSize($expectedCount, $haystack, $message = '')
+function assertNotSameSize($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNotSameSize($expectedCount, $haystack, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotSameSize',
+      func_get_args()
+    );
 }
 
 /**
@@ -927,7 +1124,10 @@ function assertNotSameSize($expectedCount, $haystack, $message = '')
  */
 function assertNotTag($matcher, $actual, $message = '', $isHtml = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertNotTag($matcher, $actual, $message, $isHtml);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNotTag',
+      func_get_args()
+    );
 }
 
 /**
@@ -938,7 +1138,10 @@ function assertNotTag($matcher, $actual, $message = '', $isHtml = TRUE)
  */
 function assertNull($actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertNull($actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertNull',
+      func_get_args()
+    );
 }
 
 /**
@@ -951,7 +1154,10 @@ function assertNull($actual, $message = '')
  */
 function assertObjectHasAttribute($attributeName, $object, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertObjectHasAttribute($attributeName, $object, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertObjectHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -964,7 +1170,10 @@ function assertObjectHasAttribute($attributeName, $object, $message = '')
  */
 function assertObjectNotHasAttribute($attributeName, $object, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertObjectNotHasAttribute($attributeName, $object, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertObjectNotHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -976,7 +1185,10 @@ function assertObjectNotHasAttribute($attributeName, $object, $message = '')
  */
 function assertRegExp($pattern, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertRegExp($pattern, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertRegExp',
+      func_get_args()
+    );
 }
 
 /**
@@ -990,20 +1202,26 @@ function assertRegExp($pattern, $string, $message = '')
  */
 function assertSame($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertSame($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertSame',
+      func_get_args()
+    );
 }
 
 /**
  * Assert that the size of two arrays (or `Countable` or `Iterator` objects)
  * is the same.
  *
- * @param integer $expected
- * @param mixed   $actual
- * @param string  $message
+ * @param array|Countable|Iterator $expected
+ * @param array|Countable|Iterator $actual
+ * @param string $message
  */
 function assertSameSize($expected, $actual, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertSameSize($expected, $actual, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertSameSize',
+      func_get_args()
+    );
 }
 
 /**
@@ -1032,7 +1250,10 @@ function assertSameSize($expected, $actual, $message = '')
  */
 function assertSelectCount($selector, $count, $actual, $message = '', $isHtml = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertSelectCount($selector, $count, $actual, $message, $isHtml);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertSelectCount',
+      func_get_args()
+    );
 }
 
 /**
@@ -1051,7 +1272,10 @@ function assertSelectCount($selector, $count, $actual, $message = '', $isHtml = 
  */
 function assertSelectEquals($selector, $content, $count, $actual, $message = '', $isHtml = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertSelectEquals($selector, $content, $count, $actual, $message, $isHtml);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertSelectEquals',
+      func_get_args()
+    );
 }
 
 /**
@@ -1070,7 +1294,10 @@ function assertSelectEquals($selector, $content, $count, $actual, $message = '',
  */
 function assertSelectRegExp($selector, $pattern, $count, $actual, $message = '', $isHtml = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertSelectRegExp($selector, $pattern, $count, $actual, $message, $isHtml);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertSelectRegExp',
+      func_get_args()
+    );
 }
 
 /**
@@ -1083,7 +1310,10 @@ function assertSelectRegExp($selector, $pattern, $count, $actual, $message = '',
  */
 function assertStringEndsNotWith($suffix, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringEndsNotWith($suffix, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringEndsNotWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1096,7 +1326,10 @@ function assertStringEndsNotWith($suffix, $string, $message = '')
  */
 function assertStringEndsWith($suffix, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringEndsWith($suffix, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringEndsWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1112,7 +1345,10 @@ function assertStringEndsWith($suffix, $string, $message = '')
  */
 function assertStringEqualsFile($expectedFile, $actualString, $message = '', $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertStringEqualsFile($expectedFile, $actualString, $message, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringEqualsFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1125,7 +1361,10 @@ function assertStringEqualsFile($expectedFile, $actualString, $message = '', $ca
  */
 function assertStringMatchesFormat($format, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringMatchesFormat($format, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringMatchesFormat',
+      func_get_args()
+    );
 }
 
 /**
@@ -1138,7 +1377,10 @@ function assertStringMatchesFormat($format, $string, $message = '')
  */
 function assertStringMatchesFormatFile($formatFile, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringMatchesFormatFile($formatFile, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringMatchesFormatFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1154,7 +1396,10 @@ function assertStringMatchesFormatFile($formatFile, $string, $message = '')
  */
 function assertStringNotEqualsFile($expectedFile, $actualString, $message = '', $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::assertStringNotEqualsFile($expectedFile, $actualString, $message, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringNotEqualsFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1167,7 +1412,10 @@ function assertStringNotEqualsFile($expectedFile, $actualString, $message = '', 
  */
 function assertStringNotMatchesFormat($format, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringNotMatchesFormat($format, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringNotMatchesFormat',
+      func_get_args()
+    );
 }
 
 /**
@@ -1180,7 +1428,10 @@ function assertStringNotMatchesFormat($format, $string, $message = '')
  */
 function assertStringNotMatchesFormatFile($formatFile, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringNotMatchesFormatFile($formatFile, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringNotMatchesFormatFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1193,7 +1444,10 @@ function assertStringNotMatchesFormatFile($formatFile, $string, $message = '')
  */
 function assertStringStartsNotWith($prefix, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringStartsNotWith($prefix, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringStartsNotWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1206,7 +1460,10 @@ function assertStringStartsNotWith($prefix, $string, $message = '')
  */
 function assertStringStartsWith($prefix, $string, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertStringStartsWith($prefix, $string, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertStringStartsWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1348,7 +1605,10 @@ function assertStringStartsWith($prefix, $string, $message = '')
  */
 function assertTag($matcher, $actual, $message = '', $isHtml = TRUE)
 {
-    return PHPUnit_Framework_Assert::assertTag($matcher, $actual, $message, $isHtml);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertTag',
+      func_get_args()
+    );
 }
 
 /**
@@ -1361,7 +1621,10 @@ function assertTag($matcher, $actual, $message = '', $isHtml = TRUE)
  */
 function assertThat($value, PHPUnit_Framework_Constraint $constraint, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertThat($value, $constraint, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertThat',
+      func_get_args()
+    );
 }
 
 /**
@@ -1373,7 +1636,10 @@ function assertThat($value, PHPUnit_Framework_Constraint $constraint, $message =
  */
 function assertTrue($condition, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertTrue($condition, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertTrue',
+      func_get_args()
+    );
 }
 
 /**
@@ -1386,7 +1652,10 @@ function assertTrue($condition, $message = '')
  */
 function assertXmlFileEqualsXmlFile($expectedFile, $actualFile, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlFileEqualsXmlFile($expectedFile, $actualFile, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlFileEqualsXmlFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1399,7 +1668,10 @@ function assertXmlFileEqualsXmlFile($expectedFile, $actualFile, $message = '')
  */
 function assertXmlFileNotEqualsXmlFile($expectedFile, $actualFile, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlFileNotEqualsXmlFile($expectedFile, $actualFile, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlFileNotEqualsXmlFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1412,7 +1684,10 @@ function assertXmlFileNotEqualsXmlFile($expectedFile, $actualFile, $message = ''
  */
 function assertXmlStringEqualsXmlFile($expectedFile, $actualXml, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlStringEqualsXmlFile($expectedFile, $actualXml, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlStringEqualsXmlFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1425,7 +1700,10 @@ function assertXmlStringEqualsXmlFile($expectedFile, $actualXml, $message = '')
  */
 function assertXmlStringEqualsXmlString($expectedXml, $actualXml, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlStringEqualsXmlString($expectedXml, $actualXml, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlStringEqualsXmlString',
+      func_get_args()
+    );
 }
 
 /**
@@ -1438,7 +1716,10 @@ function assertXmlStringEqualsXmlString($expectedXml, $actualXml, $message = '')
  */
 function assertXmlStringNotEqualsXmlFile($expectedFile, $actualXml, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlFile($expectedFile, $actualXml, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlFile',
+      func_get_args()
+    );
 }
 
 /**
@@ -1451,7 +1732,10 @@ function assertXmlStringNotEqualsXmlFile($expectedFile, $actualXml, $message = '
  */
 function assertXmlStringNotEqualsXmlString($expectedXml, $actualXml, $message = '')
 {
-    return PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlString($expectedXml, $actualXml, $message);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::assertXmlStringNotEqualsXmlString',
+      func_get_args()
+    );
 }
 
 /**
@@ -1464,7 +1748,10 @@ function assertXmlStringNotEqualsXmlString($expectedXml, $actualXml, $message = 
  */
 function at($index)
 {
-    return PHPUnit_Framework_TestCase::at($index);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::at',
+      func_get_args()
+    );
 }
 
 /**
@@ -1476,7 +1763,10 @@ function at($index)
  */
 function atLeastOnce()
 {
-    return PHPUnit_Framework_TestCase::atLeastOnce();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::atLeastOnce',
+      func_get_args()
+    );
 }
 
 /**
@@ -1489,7 +1779,10 @@ function atLeastOnce()
  */
 function attribute(PHPUnit_Framework_Constraint $constraint, $attributeName)
 {
-    return PHPUnit_Framework_Assert::attribute($constraint, $attributeName);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::attribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -1508,7 +1801,24 @@ function attribute(PHPUnit_Framework_Constraint $constraint, $attributeName)
  */
 function attributeEqualTo($attributeName, $value, $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::attributeEqualTo($attributeName, $value, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::attributeEqualTo',
+      func_get_args()
+    );
+}
+
+/**
+ * Returns a PHPUnit_Framework_Constraint_Callback matcher object.
+ *
+ * @param callable $callback
+ * @return PHPUnit_Framework_Constraint_Callback
+ */
+function callback($callback)
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::callback',
+      func_get_args()
+    );
 }
 
 /**
@@ -1520,7 +1830,10 @@ function attributeEqualTo($attributeName, $value, $delta = 0, $maxDepth = 10, $c
  */
 function classHasAttribute($attributeName)
 {
-    return PHPUnit_Framework_Assert::classHasAttribute($attributeName);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::classHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -1533,7 +1846,10 @@ function classHasAttribute($attributeName)
  */
 function classHasStaticAttribute($attributeName)
 {
-    return PHPUnit_Framework_Assert::classHasStaticAttribute($attributeName);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::classHasStaticAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -1547,7 +1863,10 @@ function classHasStaticAttribute($attributeName)
  */
 function contains($value, $checkForObjectIdentity = TRUE)
 {
-    return PHPUnit_Framework_Assert::contains($value, $checkForObjectIdentity);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::contains',
+      func_get_args()
+    );
 }
 
 /**
@@ -1560,7 +1879,25 @@ function contains($value, $checkForObjectIdentity = TRUE)
  */
 function containsOnly($type)
 {
-    return PHPUnit_Framework_Assert::containsOnly($type);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::containsOnly',
+      func_get_args()
+    );
+}
+
+/**
+ * Returns a PHPUnit_Framework_Constraint_TraversableContainsOnly matcher
+ * object.
+ *
+ * @param string $classname
+ * @return PHPUnit_Framework_Constraint_TraversableContainsOnly
+ */
+function containsOnlyInstancesOf($classname)
+{
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::containsOnlyInstancesOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -1576,7 +1913,10 @@ function containsOnly($type)
  */
 function equalTo($value, $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ignoreCase = FALSE)
 {
-    return PHPUnit_Framework_Assert::equalTo($value, $delta, $maxDepth, $canonicalize, $ignoreCase);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::equalTo',
+      func_get_args()
+    );
 }
 
 /**
@@ -1589,7 +1929,10 @@ function equalTo($value, $delta = 0, $maxDepth = 10, $canonicalize = FALSE, $ign
  */
 function exactly($count)
 {
-    return PHPUnit_Framework_TestCase::exactly($count);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::exactly',
+      func_get_args()
+    );
 }
 
 /**
@@ -1600,7 +1943,10 @@ function exactly($count)
  */
 function fileExists()
 {
-    return PHPUnit_Framework_Assert::fileExists();
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::fileExists',
+      func_get_args()
+    );
 }
 
 /**
@@ -1612,7 +1958,10 @@ function fileExists()
  */
 function greaterThan($value)
 {
-    return PHPUnit_Framework_Assert::greaterThan($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::greaterThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -1626,7 +1975,10 @@ function greaterThan($value)
  */
 function greaterThanOrEqual($value)
 {
-    return PHPUnit_Framework_Assert::greaterThanOrEqual($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::greaterThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -1638,7 +1990,10 @@ function greaterThanOrEqual($value)
  */
 function identicalTo($value)
 {
-    return PHPUnit_Framework_Assert::identicalTo($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::identicalTo',
+      func_get_args()
+    );
 }
 
 /**
@@ -1649,7 +2004,10 @@ function identicalTo($value)
  */
 function isEmpty()
 {
-    return PHPUnit_Framework_Assert::isEmpty();
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isEmpty',
+      func_get_args()
+    );
 }
 
 /**
@@ -1660,7 +2018,10 @@ function isEmpty()
  */
 function isFalse()
 {
-    return PHPUnit_Framework_Assert::isFalse();
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isFalse',
+      func_get_args()
+    );
 }
 
 /**
@@ -1672,7 +2033,10 @@ function isFalse()
  */
 function isInstanceOf($className)
 {
-    return PHPUnit_Framework_Assert::isInstanceOf($className);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isInstanceOf',
+      func_get_args()
+    );
 }
 
 /**
@@ -1683,7 +2047,10 @@ function isInstanceOf($className)
  */
 function isNull()
 {
-    return PHPUnit_Framework_Assert::isNull();
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isNull',
+      func_get_args()
+    );
 }
 
 /**
@@ -1694,18 +2061,10 @@ function isNull()
  */
 function isTrue()
 {
-    return PHPUnit_Framework_Assert::isTrue();
-}
-
-/**
- * Returns a PHPUnit_Framework_Constraint_Callback matcher object.
- *
- * @param callable $callback
- * @return PHPUnit_Framework_Constraint_Callback
- */
-function callback($callback)
-{
-    return PHPUnit_Framework_Assert::callback($callback);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isTrue',
+      func_get_args()
+    );
 }
 
 /**
@@ -1717,7 +2076,10 @@ function callback($callback)
  */
 function isType($type)
 {
-    return PHPUnit_Framework_Assert::isType($type);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::isType',
+      func_get_args()
+    );
 }
 
 /**
@@ -1729,7 +2091,10 @@ function isType($type)
  */
 function lessThan($value)
 {
-    return PHPUnit_Framework_Assert::lessThan($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::lessThan',
+      func_get_args()
+    );
 }
 
 /**
@@ -1743,7 +2108,10 @@ function lessThan($value)
  */
 function lessThanOrEqual($value)
 {
-    return PHPUnit_Framework_Assert::lessThanOrEqual($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::lessThanOrEqual',
+      func_get_args()
+    );
 }
 
 /**
@@ -1754,8 +2122,10 @@ function lessThanOrEqual($value)
  */
 function logicalAnd()
 {
-    $args = func_get_args();
-    return call_user_func_array('PHPUnit_Framework_Assert::logicalAnd', $args);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::logicalAnd',
+      func_get_args()
+    );
 }
 
 /**
@@ -1767,7 +2137,10 @@ function logicalAnd()
  */
 function logicalNot(PHPUnit_Framework_Constraint $constraint)
 {
-    return PHPUnit_Framework_Assert::logicalNot($constraint);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::logicalNot',
+      func_get_args()
+    );
 }
 
 /**
@@ -1778,8 +2151,10 @@ function logicalNot(PHPUnit_Framework_Constraint $constraint)
  */
 function logicalOr()
 {
-    $args = func_get_args();
-    return call_user_func_array('PHPUnit_Framework_Assert::logicalOr', $args);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::logicalOr',
+      func_get_args()
+    );
 }
 
 /**
@@ -1790,8 +2165,10 @@ function logicalOr()
  */
 function logicalXor()
 {
-    $args = func_get_args();
-    return call_user_func_array('PHPUnit_Framework_Assert::logicalXor', $args);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::logicalXor',
+      func_get_args()
+    );
 }
 
 /**
@@ -1803,7 +2180,10 @@ function logicalXor()
  */
 function matches($string)
 {
-    return PHPUnit_Framework_Assert::matches($string);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::matches',
+      func_get_args()
+    );
 }
 
 /**
@@ -1815,7 +2195,10 @@ function matches($string)
  */
 function matchesRegularExpression($pattern)
 {
-    return PHPUnit_Framework_Assert::matchesRegularExpression($pattern);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::matchesRegularExpression',
+      func_get_args()
+    );
 }
 
 /**
@@ -1827,7 +2210,10 @@ function matchesRegularExpression($pattern)
  */
 function never()
 {
-    return PHPUnit_Framework_TestCase::never();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::never',
+      func_get_args()
+    );
 }
 
 /**
@@ -1839,7 +2225,10 @@ function never()
  */
 function objectHasAttribute($attributeName)
 {
-    return PHPUnit_Framework_Assert::objectHasAttribute($attributeName);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::objectHasAttribute',
+      func_get_args()
+    );
 }
 
 /**
@@ -1851,7 +2240,10 @@ function objectHasAttribute($attributeName)
  */
 function onConsecutiveCalls()
 {
-    return PHPUnit_Framework_TestCase::onConsecutiveCalls();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::onConsecutiveCalls',
+      func_get_args()
+    );
 }
 
 /**
@@ -1863,7 +2255,10 @@ function onConsecutiveCalls()
  */
 function once()
 {
-    return PHPUnit_Framework_TestCase::once();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::once',
+      func_get_args()
+    );
 }
 
 /**
@@ -1875,7 +2270,10 @@ function once()
  */
 function returnArgument($argumentIndex)
 {
-    return PHPUnit_Framework_TestCase::returnArgument($argumentIndex);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::returnArgument',
+      func_get_args()
+    );
 }
 
 /**
@@ -1887,7 +2285,10 @@ function returnArgument($argumentIndex)
  */
 function returnCallback($callback)
 {
-    return PHPUnit_Framework_TestCase::returnCallback($callback);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::returnCallback',
+      func_get_args()
+    );
 }
 
 /**
@@ -1900,7 +2301,10 @@ function returnCallback($callback)
  */
 function returnSelf()
 {
-    return PHPUnit_Framework_TestCase::returnSelf();
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::returnSelf',
+      func_get_args()
+    );
 }
 
 /**
@@ -1912,7 +2316,10 @@ function returnSelf()
  */
 function returnValue($value)
 {
-    return PHPUnit_Framework_TestCase::returnValue($value);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::returnValue',
+      func_get_args()
+    );
 }
 
 /**
@@ -1924,7 +2331,10 @@ function returnValue($value)
  */
 function returnValueMap(array $valueMap)
 {
-    return PHPUnit_Framework_TestCase::returnValueMap($valueMap);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::returnValueMap',
+      func_get_args()
+    );
 }
 
 /**
@@ -1937,7 +2347,10 @@ function returnValueMap(array $valueMap)
  */
 function stringContains($string, $case = TRUE)
 {
-    return PHPUnit_Framework_Assert::stringContains($string, $case);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::stringContains',
+      func_get_args()
+    );
 }
 
 /**
@@ -1949,7 +2362,10 @@ function stringContains($string, $case = TRUE)
  */
 function stringEndsWith($suffix)
 {
-    return PHPUnit_Framework_Assert::stringEndsWith($suffix);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::stringEndsWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1961,7 +2377,10 @@ function stringEndsWith($suffix)
  */
 function stringStartsWith($prefix)
 {
-    return PHPUnit_Framework_Assert::stringStartsWith($prefix);
+    return call_user_func_array(
+      'PHPUnit_Framework_Assert::stringStartsWith',
+      func_get_args()
+    );
 }
 
 /**
@@ -1973,5 +2392,8 @@ function stringStartsWith($prefix)
  */
 function throwException(Exception $exception)
 {
-    return PHPUnit_Framework_TestCase::throwException($exception);
+    return call_user_func_array(
+      'PHPUnit_Framework_TestCase::throwException',
+      func_get_args()
+    );
 }